device: make NMDeviceFactory a class instead of an interface
An interface would make sense to allow the actual device-factory to inherit
from another type.

However, in reality interfaces (in glib) make code only more complicated and
less efficient. The device factory shall be a very simple type which describes
the devices that are supported and creates them. There is no need to make
this an interface implementation, instead just let the factories inherit the
NM_TYPE_DEVICE_FACTORY directly.